refactor: rename `TestServerTransport` to `Transport`
