+++ README.md
@@ -8,8 +8,7 @@
   [![docs](https://docs.rs/axum-test/badge.svg)](https://docs.rs/axum-test)
 </div>
 
-This is a project to make it easier to test applications built using Axum,
-using full E2E tests.
+Easy testing for applications built using Axum.
 
 ## Examples
 
@@ -17,8 +16,8 @@ You can find a thorough example in the [/examples folder](/examples/example-todo
 
 ## Features
 
-You can start a webserver running your application, query it with requests,
-and then assert the responses returned:
+Using this library, you can host your application and query against it with requests.
+Then decode the responses, and assert what is returned:
 
 ```rust
   use ::axum::Router;
@@ -36,7 +35,7 @@ and then assert the responses returned:
       let app = Router::new()
           .route("/ping", get(get_ping));
 
-      // Run the server on a random address.
+      // Run the application for testing.
       let server = TestServer::new(app).unwrap();
 
       // Get the request.
@@ -48,10 +47,11 @@ and then assert the responses returned:
   }
 ```
 
-The `TestServer` will start on a random port, allowing multiple servers to run in parallel.
-That allows tests to be run in parallel.
+The `TestServer` can run requests directly against your application with a mocked network,
+or the application can run on a random port (with real network reqeusts being made).
+In both cases allowing multiple servers to run in parallel, across your tests.
 
-This behaviour can be changed in the `TestServerConfig`, by selecting a custom ip or port to always use.
+This behaviour can be changed in the `TestServerConfig`, by selecting the `transport` to be used.
 
 ### Request building
 
@@ -63,11 +63,11 @@ Querying your application on the `TestServer` supports all of the common request
  - Status code reading and assertions
  - Assertions for defining what you expect to have returned
 
-### Remembers cookies across requests
+As well as also including:
 
-`axum-test` supports preserving Cookies from responses,
-for use in follow up requests to same `TestServer`.
-This is similar to how a browser will store cookies from requests,
-and can help with tests where you need to authenticate first.
+ - Saving cookies returned for use across future requests.
+ - Setting headers and query parameters for use across all TestRequests.
+ - Can optionally run requests using a real web server.
+ - Automatic status assertions for checking requests always succeed or fail.
 
-This is _off_ by default, and can be enabled in the `TestServerConfig` by setting `save_cookies` to true.
+ and much more!
