+++ src/internals/transport_layer/mock_transport_layer.rs
@@ -1,18 +1,17 @@
 use ::anyhow::Error as AnyhowError;
 use ::anyhow::Result;
 use ::async_trait::async_trait;
+use ::axum::Router;
+use ::bytes::Bytes;
+use ::http::response::Parts;
 use ::http::Request;
+use ::hyper::body::to_bytes;
 use ::hyper::Body;
 use ::std::fmt::Debug;
+use ::tower::util::ServiceExt;
+use ::tower::Service;
 
-use super::TransportLayer;
-
-use axum::Router;
-use bytes::Bytes;
-use http::response::Parts;
-use hyper::body::to_bytes;
-use tower::util::ServiceExt;
-use tower::Service;
+use crate::transport_layer::TransportLayer;
 
 pub struct MockTransportLayer<S> {
     service: S,
