+++ src/internals/transport_layer/mod.rs
@@ -3,6 +3,3 @@ pub use self::http_transport_layer::*;
 
 mod mock_transport_layer;
 pub use self::mock_transport_layer::*;
-
-mod transport_layer;
-pub use self::transport_layer::*;
