+++ src/test_server/inner_mock_server.rs
@@ -4,8 +4,8 @@ use ::std::sync::Mutex;
 use ::url::Url;
 
 use super::InnerServer;
-use crate::internals::TransportLayer;
-use crate::transport::IntoMockTransportLayer;
+use crate::transport_layer::IntoMockTransportLayer;
+use crate::transport_layer::TransportLayer;
 
 pub struct InnerMockServer {
     transport: Arc<Mutex<Box<dyn TransportLayer>>>,
@@ -23,12 +23,9 @@ impl InnerMockServer {
 }
 
 impl InnerServer for InnerMockServer {
-    /// Returns the local web address for the test server.
-    ///
-    /// By default this will be something like `http://0.0.0.0:1234/`,
-    /// where `1234` is a randomly assigned port numbr.
+    /// Returns 'example.com'.
     fn server_address<'a>(&'a self) -> &'a str {
-        &""
+        &"http://example.com"
     }
 
     fn url(&self) -> Url {
