+++ src/transport_layer/into_mock_transport_layer.rs
@@ -4,7 +4,7 @@ use ::axum::Router;
 use ::hyper::server::conn::AddrStream;
 
 use crate::internals::MockTransportLayer;
-use crate::internals::TransportLayer;
+use crate::transport_layer::TransportLayer;
 
 /// This exists to gloss over the differences between Axum's
 /// [`IntoMakeService`](::axum::routing::IntoMakeService) and [`IntoMakeServiceWithConnectInfo`](::axum::extract::connect_info::IntoMakeServiceWithConnectInfo) types.
@@ -46,7 +46,7 @@ mod test_into_mock_transport_layer_for_router {
 
     use crate::TestServer;
     use crate::TestServerConfig;
-    use crate::TestServerTransport;
+    use crate::Transport;
 
     async fn get_ping() -> &'static str {
         "pong!"
@@ -63,7 +63,7 @@ mod test_into_mock_transport_layer_for_router {
 
         // Run the server.
         let config = TestServerConfig {
-            transport: TestServerTransport::MockHttp,
+            transport: Transport::MockHttp,
             ..TestServerConfig::default()
         };
         let server = TestServer::new_with_config(app, config).expect("Should create test server");
@@ -81,7 +81,7 @@ mod test_into_mock_transport_layer_for_router {
 
         // Run the server.
         let config = TestServerConfig {
-            transport: TestServerTransport::MockHttp,
+            transport: Transport::MockHttp,
             ..TestServerConfig::default()
         };
         let server = TestServer::new_with_config(app, config).expect("Should create test server");
@@ -100,7 +100,7 @@ mod test_into_mock_transport_layer_for_into_make_service {
 
     use crate::TestServer;
     use crate::TestServerConfig;
-    use crate::TestServerTransport;
+    use crate::Transport;
 
     async fn get_ping() -> &'static str {
         "pong!"
@@ -119,7 +119,7 @@ mod test_into_mock_transport_layer_for_into_make_service {
 
         // Run the server.
         let config = TestServerConfig {
-            transport: TestServerTransport::MockHttp,
+            transport: Transport::MockHttp,
             ..TestServerConfig::default()
         };
         let server = TestServer::new_with_config(app, config).expect("Should create test server");
@@ -138,7 +138,7 @@ mod test_into_mock_transport_layer_for_into_make_service {
 
         // Run the server.
         let config = TestServerConfig {
-            transport: TestServerTransport::MockHttp,
+            transport: Transport::MockHttp,
             ..TestServerConfig::default()
         };
         let server = TestServer::new_with_config(app, config).expect("Should create test server");
@@ -156,7 +156,7 @@ mod test_into_mock_transport_layer_for_into_make_service_with_connect_info {
 
     use crate::TestServer;
     use crate::TestServerConfig;
-    use crate::TestServerTransport;
+    use crate::Transport;
 
     async fn get_ping() -> &'static str {
         "pong!"
@@ -172,7 +172,7 @@ mod test_into_mock_transport_layer_for_into_make_service_with_connect_info {
 
         // Build the server.
         let config = TestServerConfig {
-            transport: TestServerTransport::MockHttp,
+            transport: Transport::MockHttp,
             ..TestServerConfig::default()
         };
         TestServer::new_with_config(app, config).expect("Should create test server");
