+++ src/transport_layer/mod.rs
@@ -3,3 +3,6 @@ pub use self::into_http_transport_layer::*;
 
 mod into_mock_transport_layer;
 pub use self::into_mock_transport_layer::*;
+
+mod transport_layer;
+pub use self::transport_layer::*;
